Introduce .advanceOver(token)
We used .accept(token) in a few places, but this was wrong, because
.accept(token) doesn't give an error if the token kind does not
match. This new method is like .expect(token) and then advance the
lexer. The fact that all the tests passed with this being wrong
means that there are some parse failure paths the tests do not
cover.